+++ epserde/src/impls/prim.rs
@@ -74,7 +74,7 @@ macro_rules! impl_prim_ser_des {
             }
         }
 
-		unsafe impl DeserializeInner for $ty {
+		impl DeserializeInner for $ty {
             #[inline(always)]
             unsafe fn _deserialize_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<$ty> {
                 let mut buf = [0; size_of::<$ty>()];
@@ -116,7 +116,7 @@ macro_rules! impl_nonzero_ser_des {
             }
         }
 
-		unsafe impl DeserializeInner for $ty {
+		impl DeserializeInner for $ty {
             #[inline(always)]
             unsafe fn _deserialize_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<$ty> {
                 let mut buf = [0; size_of::<$ty>()];
@@ -186,7 +186,7 @@ impl SerializeInner for bool {
     }
 }
 
-unsafe impl DeserializeInner for bool {
+impl DeserializeInner for bool {
     #[inline(always)]
     unsafe fn _deserialize_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<bool> {
         Ok(u8::_deserialize_full_inner(backend)? != 0)
@@ -215,7 +215,7 @@ impl SerializeInner for char {
     }
 }
 
-unsafe impl DeserializeInner for char {
+impl DeserializeInner for char {
     #[inline(always)]
     unsafe fn _deserialize_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<Self> {
         Ok(char::from_u32(u32::_deserialize_full_inner(backend)?).unwrap())
@@ -242,7 +242,7 @@ impl SerializeInner for () {
     }
 }
 
-unsafe impl DeserializeInner for () {
+impl DeserializeInner for () {
     #[inline(always)]
     unsafe fn _deserialize_full_inner(_backend: &mut impl ReadWithPos) -> deser::Result<Self> {
         Ok(())
@@ -295,7 +295,7 @@ impl<T: ?Sized> SerializeInner for PhantomData<T> {
     }
 }
 
-unsafe impl<T: ?Sized> DeserializeInner for PhantomData<T> {
+impl<T: ?Sized> DeserializeInner for PhantomData<T> {
     #[inline(always)]
     unsafe fn _deserialize_full_inner(_backend: &mut impl ReadWithPos) -> deser::Result<Self> {
         Ok(PhantomData::<T>)
@@ -346,7 +346,7 @@ impl<T: SerializeInner + TypeHash + AlignHash> SerializeInner for Option<T> {
     }
 }
 
-unsafe impl<T: DeserializeInner> DeserializeInner for Option<T> {
+impl<T: DeserializeInner> DeserializeInner for Option<T> {
     #[inline(always)]
     unsafe fn _deserialize_full_inner(backend: &mut impl ReadWithPos) -> deser::Result<Self> {
         let tag = u8::_deserialize_full_inner(backend)?;